Group Class, Service, Repository, Controller

+++ src/main/java/org/Aruiz/repository/GroupRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Long> {
-    // Métodos personalizados si es necesario
+public interface GroupRepository extends JpaRepository<Group, Integer> {
+
 }
